api/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
PyMySQL==1.1.0
python-ldap==3.4.3
ldap3>=2.9.1